+++ src/stringifiers/mysql.js
@@ -35,7 +35,7 @@ async function _stringifySqlAST(parent, node, prefix, context, selections, joins
     // generate the join or joins
     // this condition is for single joins (one-to-one or one-to-many relations)
     if (node.sqlJoin) {
-      const joinCondition = node.sqlJoin(`${quote(parent.as)}`, `${quote(node.as)}`, node.args || {})
+      const joinCondition = await node.sqlJoin(`${quote(parent.as)}`, `${quote(node.as)}`, node.args || {}, context)
 
       joins.push(
         `LEFT JOIN ${node.name} AS ${quote(node.as)} ON ${joinCondition}`
@@ -43,8 +43,8 @@ async function _stringifySqlAST(parent, node, prefix, context, selections, joins
     // this condition is through a join table (many-to-many relations)
     } else if (node.joinTable) {
       if (!node.sqlJoins) throw new Error('Must set "sqlJoins" for a join table.')
-      const joinCondition1 = node.sqlJoins[0](`${quote(parent.as)}`, `${quote(node.joinTableAs)}`, node.args || {})
-      const joinCondition2 = node.sqlJoins[1](`${quote(node.joinTableAs)}`, `${quote(node.as)}`, node.args || {})
+      const joinCondition1 = await node.sqlJoins[0](`${quote(parent.as)}`, `${quote(node.joinTableAs)}`, node.args || {}, context)
+      const joinCondition2 = await node.sqlJoins[1](`${quote(node.joinTableAs)}`, `${quote(node.as)}`, node.args || {}, context)
 
       joins.push(
         `LEFT JOIN ${node.joinTable} AS ${quote(node.joinTableAs)} ON ${joinCondition1}`,
@@ -59,7 +59,7 @@ async function _stringifySqlAST(parent, node, prefix, context, selections, joins
 
     // recurse thru nodes
     for (let child of node.children) {
-      _stringifySqlAST(node, child, [ ...prefix, node.as ], context, selections, joins, wheres)
+      await _stringifySqlAST(node, child, [ ...prefix, node.as ], context, selections, joins, wheres)
     }
 
     break